third_party/blink/web_tests/external/wpt/css/cssom-view/elementFromPosition-expected.txt

This is a testharness.js-based test.
[FAIL] test some point of the element: top right corner
  assert_equals: expected true but got false
[FAIL] test some point of the element: right line
  assert_equals: expected true but got false
[FAIL] test some point of the element: bottom left corner
  assert_equals: expected true but got false
[FAIL] test some point of the element: bottom line
  assert_equals: expected true but got false
[FAIL] test some point of the element: bottom right corner
  assert_equals: expected true but got false
Harness: the test ran to completion.
